requirements.txt
numpy==1.24.1
opencv_python
pandas==1.3.4
Pillow==9.4.0
scikit_learn==1.2.2
scipy==1.9.1
scikit-image==0.20.0
tabulate==0.9.0
tqdm==4.65.0
diffusers
matplotlib==3.7.1
timm==0.4.12
transformers
accelerate